backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.9
pydantic-settings>=2.5
httpx>=0.27
asyncpg>=0.30
sqlalchemy[asyncio]>=2.0
alembic>=1.13
python-jose[cryptography]>=3.3
bcrypt>=4.0,<5.0
python-multipart>=0.0.9
structlog>=24.1
slowapi>=0.1.9
